src/requirements.txt
# Flask
fastapi
uvicorn
loguru
# uwsgi
gunicorn
gevent
ansible
mysql-connector
paramiko
netmiko
python-crontab
requests
pysocks
ruamel.yaml
werkzeug

mycli
pymysql
